+++ components/Gallery/GalleryElement.vue
