+++ components/Home/GalleryLinkWarning.vue
