+++ components/shared/ContentIndicator.vue
